trying to implement new API

+++ teste-pag-ios/Model/Movie.swift
@@ -12,7 +12,7 @@ struct Movie: Codable {
     let vote_count: Int
     let id: Int
     let video: Bool
-    let vote_average: Int
+    let vote_average: Double
     let title: String
     let popularity: Double
     let poster_path: String
